v1.3.4
API 15
Enabled use within Field Operations.

+++ FastWaymarksPlugin/Plugin.cs
@@ -111,7 +111,7 @@ public sealed class Plugin : IDalamudPlugin
         ToggleMainUI();
     }
 
-    private void TerritoryChanged(ushort territoryType)
+    private void TerritoryChanged(uint territoryType)
     {
         Plugin.Log.Debug($"Territory Changed to: {territoryType}");
         MainWindow.UpdateMapID();
